+++ inst/examples/Ex_load_program_preload.R
@@ -1,0 +1,9 @@
+############################ Start of load_program_preload example ########################
+
+manifest <- read_program_preload_manifest(source_package = "opencltools")
+print(manifest)
+
+preload <- load_program_preload(source_package = "opencltools")
+cat("Preload bytes:", attr(preload, "nbytes_concatenated"), "\n")
+
+## End of load_program_preload example